new validations

+++ src/modules/followers/services/followers.service.ts
@@ -15,9 +15,12 @@ export class FollowersService {
   // Function to create a new follower
   async createFollower(createFollowerDto: CreateFollowerDto): Promise<FollowersEntity> { 
     try{
+      if(!createFollowerDto) {
+        throw new Error('please provide all fields');
+      };
       const follower =  this.followerRepository.create(createFollowerDto);
       if(!follower) {
-        throw new Error('Follower not created, please give all required fields');
+        throw new Error('Follower not created error from createFollower');
       };
   
       return await this.followerRepository.save(follower);
@@ -30,6 +33,9 @@ export class FollowersService {
   // Function to find all followings by follower ID
   async findFollowingsById(followerId: FollowersEntity['follower']): Promise<String[]> {
     try{
+      if(!followerId) {
+        throw new Error('please provide all fields');
+      };
       const followings = await this.followerRepository.find({ where: { follower: followerId } });
     
       if (!followings || followings.length === 0) {
@@ -48,6 +54,9 @@ export class FollowersService {
   // Function to find followers by user ID
   async findFollowersByUser(followingID: FollowersEntity['following']): Promise<String[]> { 
     try{
+      if(!followingID) {
+        throw new Error('please provide all fields');
+      };
       const followers = await this.followerRepository.find({ where: { following: followingID } });
       if(!followers) {
         throw new Error('No followers found for the provided userId');
@@ -65,17 +74,19 @@ export class FollowersService {
   // Function to delete a follower by ID
   async deleteFollower(followerId: string): Promise<string> { 
     try{
-          // find follower by ID
-    const follower = await this.followerRepository.findOneBy({id: followerId});
-    if(!follower) {
-      throw new Error('Follower not found');
-    };
-    // delete follower
-    const deleted = await this.followerRepository.remove(follower);
-    if(!deleted) {
-      throw new Error('Follower not deleted');
-    };
-    return 'Follower deleted';
+      if(!followerId) {
+        throw new Error('please provide all fields');
+      };
+      const follower = await this.followerRepository.findOneBy({id: followerId});
+      if(!follower) {
+        throw new Error('Follower not found');
+      };
+      // delete follower
+      const deleted = await this.followerRepository.remove(follower);
+      if(!deleted) {
+        throw new Error('Follower not deleted');
+      };
+      return 'Follower deleted';
     }
     catch(err){
       throw new Error(err);
